Fix compiler warning: sketch/Dcodes.cpp: In function 'void dcode_3()':
sketch/Dcodes.cpp:117:24: warning: comparison between signed and unsigned integer expressions [-Wsign-compare]
    for (int i = 0; i < count; i++)

+++ Firmware/Dcodes.cpp
@@ -114,7 +114,7 @@ void dcode_3()
 		count = parse_hex(strchr_pointer + 1, data, 16);
 		if (count > 0)
 		{
-			for (int i = 0; i < count; i++)
+			for (uint16_t i = 0; i < count; i++)
 				eeprom_write_byte((uint8_t*)(address + i), data[i]);
 			printf_P(_N("%d bytes written to EEPROM at address 0x%04x"), count, address);
 			putchar('\n');
